+++ src/Windows/WindowsSocket.cpp
@@ -145,7 +145,7 @@ namespace es
         return accepting_socket;
     }
 
-    ssize_t WindowsSocket::receive_data(char* buffer, int buffer_size)
+    int64_t WindowsSocket::receive_data(char* buffer, int buffer_size)
     {
         int bytes = recv(m_socket, buffer, buffer_size, 0);
 
@@ -155,7 +155,7 @@ namespace es
         return bytes;
     }
 
-    ssize_t WindowsSocket::receive_data_from(char* buffer, int buffer_size, EndPoint& sender_end_point)
+    int64_t WindowsSocket::receive_data_from(char* buffer, int buffer_size, EndPoint& sender_end_point)
     {
         sockaddr_storage sender{};
         int sender_length = sizeof(sender);
@@ -187,7 +187,7 @@ namespace es
         return bytes;
     }
 
-    ssize_t WindowsSocket::send_data(const char* buffer, int buffer_size)
+    int64_t WindowsSocket::send_data(const char* buffer, int buffer_size)
     {
         int bytes = send(m_socket, buffer, buffer_size, 0);
